+++ NppCSharpPluginPack/Forms/DarkModeTestForm.cs
@@ -4,17 +4,15 @@ using NppDemo.Utils;
 
 namespace NppDemo.Forms
 {
-    public partial class DarkModeTestForm : Form
+    public partial class DarkModeTestForm : FormBase
     {
         private SelectionRememberingForm selectionRememberingForm;
 
-        public DarkModeTestForm(SelectionRememberingForm selectionRememberingForm)
+        public DarkModeTestForm(SelectionRememberingForm selectionRememberingForm) : base(false, false)
         {
             InitializeComponent();
-            NppFormHelper.RegisterFormIfModeless(this, false);
             this.selectionRememberingForm = selectionRememberingForm;
             selectionRememberingForm.AddOwnedForm(this);
-            FormStyle.ApplyStyle(this, Main.settings.use_npp_styling);
             comboBox1.SelectedIndex = 0;
             DataGridViewRow row = new DataGridViewRow();
             row.CreateCells(dataGridView1);