+++ NppCSharpPluginPack/Forms/FormBase.Designer.cs
@@ -1,0 +1,48 @@
+﻿namespace NppDemo.Forms
+{
+    partial class FormBase
+    {
+        /// <summary>
+        /// Required designer variable.
+        /// </summary>
+        private System.ComponentModel.IContainer components = null;
+
+        /// <summary>
+        /// Clean up any resources being used.
+        /// </summary>
+        /// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing && (components != null))
+            {
+                NppFormHelper.UnregisterFormIfModeless(this, IsModal);
+                components.Dispose();
+            }
+            base.Dispose(disposing);
+        }
+
+        #region Windows Form Designer generated code
+
+        /// <summary>
+        /// Required method for Designer support - do not modify
+        /// the contents of this method with the code editor.
+        /// </summary>
+        private void InitializeComponent()
+        {
+            this.SuspendLayout();
+            // 
+            // FormBase
+            // 
+            this.AutoScaleDimensions = new System.Drawing.SizeF(8F, 16F);
+            this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+            this.ClientSize = new System.Drawing.Size(800, 450);
+            this.Name = "FormBase";
+            this.Text = "DockingFormBase";
+            this.VisibleChanged += new System.EventHandler(this.FormBase_VisibleChanged);
+            this.ResumeLayout(false);
+
+        }
+
+        #endregion
+    }
+}