+++ NppCSharpPluginPack/Forms/PopupDialog.cs
@@ -1,17 +1,13 @@
-﻿using NppDemo.Utils;
-using Kbg.NppPluginNET;
-using System.Windows.Forms;
+﻿using System.Windows.Forms;
 using ExampleDependency;
 
 namespace NppDemo.Forms
 {
-    public partial class PopupDialog : Form
+    public partial class PopupDialog : FormBase
     {
-        public PopupDialog()
+        public PopupDialog() : base(true, false)
         {
             InitializeComponent();
-            NppFormHelper.RegisterFormIfModeless(this, true);
-            FormStyle.ApplyStyle(this, Main.settings.use_npp_styling);
             ComboBox1.Enabled = ComboBox1EnabledCheckBox.Checked;
         }
 
@@ -28,23 +24,6 @@ namespace NppDemo.Forms
                 : "ComboBox1 is disabled";
             var exampleClassMember = new ExampleClass(msg);
             MessageBox.Show(exampleClassMember.ToString());
-        }
-
-        /// <summary>
-        /// suppress the default response to the Tab key
-        /// </summary>
-        /// <param name="keyData"></param>
-        /// <returns></returns>
-        protected override bool ProcessDialogKey(Keys keyData)
-        {
-            if (keyData.HasFlag(Keys.Tab)) // this covers Tab with or without modifiers
-                return true;
-            return base.ProcessDialogKey(keyData);
-        }
-
-        private void PopupDialog_KeyUp(object sender, KeyEventArgs e)
-        {
-            NppFormHelper.GenericKeyUpHandler(this, sender, e, true);
         }
     }
 }